requirements.txt
discord.py==0.14.3
Pyrebase==3.0.24
